Remove shebang flag file path relativization
`deno run` flags that reference file paths will now be relative to the
`cwd` (i.e. the root of the project) instead of being relative to the
entrypoint file.

This was done to simplify the code base, as well as remove the weird
dichotomy of the referenced files being relative to the endpoint, but
during runtime the file operations are relative to cwd. This was
confusing.